ChangesModel/Keyindicator/MemberPage/MemberPageRefData/decimaltoLong

+++ BOTS_BL/Models/KeyIndecators/MemberPage.cs
@@ -29,14 +29,14 @@ namespace BOTS_BL.Models
         public string MemberName { get; set; }
         public long? ReferralGenerated { get; set; }
         public long? ReferralTransacted { get; set; }
-        public decimal? BusinessGenerated { get; set; }
+        public long? BusinessGenerated { get; set; }
         public string ReferralMobileNo { get; set; }
         public string ReferralName { get; set; }
         public int? TransactionCount { get; set; }
         public string Source { get; set; }
         public string SourceDesc { get; set; }
         public int? RegCount { get; set; }
-        public decimal? Percentage { get; set; }
+        public long? Percentage { get; set; }
 
 
     }
